refactor: rename/moves so it match organisation on ee
