+++ botfront/imports/api/graphql/index.js
@@ -1,9 +1,9 @@
 import path from 'path';
 import { fileLoader, mergeTypes } from 'merge-graphql-schemas';
 
-import entityDistributionResolver from './nlu/entityDistributionResolver';
-import intentDistributionResolver from './nlu/intentDistributionResolver';
-import conversationsResolver from './conversations/conversationsResolver';
+import entityDistributionResolver from './nlu/resolvers/entityDistributionResolver';
+import intentDistributionResolver from './nlu/resolvers/intentDistributionResolver';
+import conversationsResolver from './conversations/resolvers/conversationsResolver';
 /*
 * mergeResolvers doesnt work (https://github.com/Urigo/merge-graphql-schemas/issues?utf8=%E2%9C%93&q=unexpected+token) 
 * so we need to import resolvers one by one
