+++ botfront/imports/api/graphql/nlu/resolvers/intentDistributionResolver.js
@@ -1,10 +1,9 @@
-import { getIntentDistribution } from '../nlu.analytics';
+import { getIntentDistribution } from '../mongo/nlu.analytics';
 
 export default {
     Query: {
         async intentDistribution(parent, args, context, info) {
             return getIntentDistribution(args.modelId);
-            // return find(authors, { id: args.id });
         },
     },
     IntentCount: {
