+++ botfront/imports/api/graphql/nlu/schemas/entityDistribution.types.graphql
