+++ botfront/imports/api/graphql/nlu/schemas/index.js
@@ -1,0 +1,7 @@
+import entityDistributionTypes from './entityDistribution.types.graphql';
+import intentDistributionTypes from './intentDistribution.types.graphql';
+
+export default [
+    entityDistributionTypes,
+    intentDistributionTypes,
+];
