+++ botfront/imports/api/graphql/nlu/schemas/intentDistribution.types.graphql
